+++ shared/api/tokenizer_impl.h
@@ -69,27 +69,27 @@ class TokenizerImpl : public OrtxObjectImpl {
   std::string chat_template;
   std::vector<std::unordered_map<std::string, std::string>> messages;
 
-  OrtxStatus PhiVisionChatTemplate(std::string* output, bool add_generation_prompt);
+  OrtxStatus PhiVisionChatTemplate(std::string& output, bool add_generation_prompt);
   
-  OrtxStatus Phi3ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token);
+  OrtxStatus Phi3ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token);
 
-  OrtxStatus Phi3SmallChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
+  OrtxStatus Phi3SmallChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
 
-  OrtxStatus Phi3MediumChatTemplate(std::string* output);
+  OrtxStatus Phi3MediumChatTemplate(std::string& output);
   
-  OrtxStatus Phi4ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token);
+  OrtxStatus Phi4ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token);
   
-  OrtxStatus Llama2ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
+  OrtxStatus Llama2ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
 
-  OrtxStatus Llama3ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
+  OrtxStatus Llama3ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
 
-  OrtxStatus Llama3_2ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::vector<std::string>& custom_tools, bool tools_in_user_message, const std::string& strftime_now, const std::string& bos_token);
+  OrtxStatus Llama3_2ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::vector<std::string>& custom_tools, bool tools_in_user_message, const std::string& strftime_now, const std::string& bos_token);
 
-  OrtxStatus Llama3_3ChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::vector<std::string>& custom_tools, const std::vector<std::string>& builtin_tools, bool tools_in_user_message, const std::string& date_string, const std::string& bos_token);
+  OrtxStatus Llama3_3ChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::vector<std::string>& custom_tools, const std::vector<std::string>& builtin_tools, bool tools_in_user_message, const std::string& date_string, const std::string& bos_token);
   
-  OrtxStatus DeepSeekChatTemplate(std::string* output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
+  OrtxStatus DeepSeekChatTemplate(std::string& output, bool add_generation_prompt, const std::string& eos_token, const std::string& bos_token);
   
-  OrtxStatus ApplyChatTemplate(std::vector<std::unordered_map<std::string, std::string>> messages, std::string* output, bool add_generation_prompt);
+  OrtxStatus ApplyChatTemplate(std::vector<std::unordered_map<std::string, std::string>> messages, std::string& output, bool add_generation_prompt);
 
   OrtxStatus Id2Token(extTokenId_t id, std::string& token, TokenizerDecodingState** state) const;
